move pages request to usePages
